Move testcase to subdir
